Ajuste de borrar V1 al momento de Guardar

+++ app/src/main/java/com/jonathanev/review/data/filesystem/GuiaRepositoryImpl.kt
@@ -187,7 +187,9 @@ class GuiaRepositoryImpl @Inject constructor(
                 return false
             }
 
-            File(currentPath).delete()
+            if (guideDomainModel.version == GuideVersion.V1){
+                File(currentPath).delete()
+            }
             true
         } catch (e: Exception) {
             false
